+++ ContentPatcher/ModEntry.cs
@@ -418,7 +418,20 @@ namespace ContentPatcher
                     fakePatchContext.SetLocalValue(ConditionType.TargetWithoutPath.ToString(), "");
                     tokenParser = new TokenParser(fakePatchContext, current.Manifest, current.Migrator, installedMods);
 
-                    // load patches
+                    // sanity check
+                    {
+                        int[] nullPositions = content.Changes
+                            .Select((patch, index) => new { patch, index })
+                            .Where(p => p.patch == null)
+                            .Select(p => p.index + 1)
+                            .ToArray();
+                        if (nullPositions.Any())
+                        {
+                            this.Monitor.Log($"Error loading content pack '{current.Manifest.Name}'. Found null patch{(nullPositions.Length == 1 ? "" : "es")} at position{(nullPositions.Length == 1 ? "" : "s")} {string.Join(", ", nullPositions)}.", LogLevel.Error);
+                            continue;
+                        }
+                    }
+
                     content.Changes = this.SplitPatches(content.Changes).ToArray();
                     this.NamePatches(current.ManagedPack, content.Changes);
                     foreach (PatchConfig patch in content.Changes)